+++ sync-for-expenses/RELEASES.md
@@ -121,3 +121,13 @@ Based on:
 - [csharp v3.3.0] sync-for-expenses
 ### Releases
 - [NuGet v3.3.0] https://www.nuget.org/packages/Codat.Sync.Expenses/3.3.0 - sync-for-expenses
+
+## 2023-11-13 01:29:20
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.0.0] sync-for-expenses
+### Releases
+- [NuGet v4.0.0] https://www.nuget.org/packages/Codat.Sync.Expenses/4.0.0 - sync-for-expenses